Removal of publishable service provider typo
Clears up a line about the publishing of the service provider. On line 351 is say it'll use `/resources/stubs/{$nameOfYourServiceProvider}.php.stub` but on line 359 it says `/resources/stubs/{$nameOfYourServiceProvider}.php`.

I've updated line 359 to use `/resources/stubs/{$nameOfYourServiceProvider}.php.stub` which is the correct value for the file the package tools will actually use.