Pest exception handling improvements
This commit replaces the Pest standard `test()` function with one which wraps the closure in another closure that rethrows any exceptions thrown during ServiceProvider register/boot methods at a point in the Pest test cycle where such exceptions can be properly tested or cause a fail.

This allows all such exceptions to be reported properly without requiring any methods to be called inside the tests themselves to rethrow any such exception, and for exceptions still to be tested.

This is completely intuitive.