Separate `hasXXXX` into `loadsXXXX` and `publishesXXXX`
One of the problems of the current approach is that whenever you load something (like a route file) it makes it publishable as well - which results in a LOT of tags in the `artisan vendor:publish` tags list.

This major commit changes splits the `hasXXXX` methods which do both into separate `loadsXXXX` and `publishesXXXX` methods so that the user can explicitly choose what to load and / or publish.

Full backwards compatibility is provided with the previous `hasXXXX` methods.

We now have 247 tests and overall code coverage of 89%.

Still to do:

1. Update Installer to improve validation and to provide publishing of all publishable functionality without having to list them explicitly.

2. Prevent InvalidPackage exceptions when AppDebug is false or Env is Production.

3. Get test coverage to 100% (or as close to it as possible - some functionality is clearly still not tested.

4. QA review of tests to ensure that all possible validations are included and that all validations are tested.

5. QA review of documentation to ensure it is fully consistent with the new functionality.